% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_se.R
\name{bootstrap_se}
\alias{bootstrap_se}
\title{Compute bootstrap-based standard error estimates for variable importance}
\usage{
bootstrap_se(Y = NULL, f1 = NULL, f2 = NULL, type = "r_squared", b = 1000)
}
\arguments{
\item{Y}{the outcome.}

\item{f1}{the fitted values from a flexible estimation technique
regressing Y on X.}

\item{f2}{the fitted values from a flexible estimation technique
regressing either (a) \code{f1} or (b) Y on X withholding the columns in
\code{indx}.}

\item{type}{the type of importance to compute; defaults to
\code{r_squared}, but other supported options are \code{auc},
\code{accuracy}, \code{deviance}, and \code{anova}.}

\item{b}{the number of bootstrap replicates (defaults to 1000)}
}
\value{
a bootstrap-based standard error estimate
}
\description{
Compute bootstrap-based standard error estimates for variable importance
}
