% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_inputs}
\alias{check_inputs}
\title{Check inputs to a call to vim, cv_vim, or sp_vim}
\usage{
check_inputs(Y, X, f1, f2, indx)
}
\arguments{
\item{Y}{the outcome}

\item{X}{the covariates}

\item{f1}{estimator of the population-optimal prediction function
using all covariates}

\item{f2}{estimator of the population-optimal prediction function
using the reduced set of covariates}

\item{indx}{the index or indices of the covariate(s) of interest}
}
\value{
None. Called for the side effect of stopping the algorithm if
  any inputs are in an unexpected format.
}
\description{
Check inputs to a call to vim, cv_vim, or sp_vim
}
\details{
Ensure that inputs to \code{vim}, \code{cv_vim}, and \code{sp_vim}
  follow the correct formats.
}
