% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_z}
\alias{create_z}
\title{Create complete-case outcome, weights, and Z}
\usage{
create_z(Y, C, Z, X, ipc_weights)
}
\arguments{
\item{Y}{the outcome}

\item{C}{indicator of missing or observed}

\item{Z}{the covariates observed in phase 1 and 2 data}

\item{X}{all covariates}

\item{ipc_weights}{the weights}
}
\value{
a list, with the complete-case outcome, weights, and Z matrix
}
\description{
Create complete-case outcome, weights, and Z
}
