% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_folds}
\alias{.make_folds}
\title{Create Folds for Cross-Fitting}
\usage{
.make_folds(y, V = 2, stratified = FALSE, probs = rep(1/V, V))
}
\arguments{
\item{y}{the outcome}

\item{V}{the number of folds}

\item{stratified}{should the folds be stratified based on the outcome?}
}
\value{
a vector of folds
}
\description{
Create Folds for Cross-Fitting
}
\keyword{internal}
