% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_predictiveness_cv.R
\name{est_predictiveness_cv}
\alias{est_predictiveness_cv}
\title{Estimate a nonparametric predictiveness functional using cross-validation}
\usage{
est_predictiveness_cv(
  fitted_values,
  y,
  folds,
  type = "r_squared",
  x = NULL,
  C = rep(1, length(y)),
  ipc_weights = rep(1, length(y)),
  ipc_fit_type = "external",
  ipc_eif_preds = rep(1, length(y)),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{fitted_values}{fitted values from a regression function; a list of length V, where each object is a set of predictions on the validation data.}

\item{y}{the outcome.}

\item{folds}{the cross-validation folds}

\item{x}{the covariates, only used if \code{ipc_weights} are entered (defaults to \code{NULL}).}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes unobserved).}

\item{ipc_weights}{weights for inverse probability of coarsening (e.g., inverse weights from a two-phase sample) weighted estimation.}

\item{ipc_fit_type}{if "external", then use \code{ipc_eif_preds}; if "SL", fit a SuperLearner to determine the correction to the efficient influence function}

\item{ipc_eif_preds}{if \code{ipc_fit_type = "external"}, the fitted values from a regression of the full-data EIF on the fully observed covariates/outcome; otherwise, not used.}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}

\item{...}{other arguments to SuperLearner, if \code{ipc_fit_type = "SL"}.}
}
\value{
The estimated measure of predictiveness.
}
\description{
Compute nonparametric estimates of the chosen measure of predictiveness.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
