% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_vimp.R
\name{est_vimp}
\alias{est_vimp}
\title{Estimate variable importance}
\usage{
est_vimp(
  full,
  reduced,
  y,
  folds,
  type = "r_squared",
  x = NULL,
  C = rep(1, length(y)),
  ipc_weights = rep(1, length(y)),
  ipc_fit_type = "external",
  ipc_eif_preds = rep(1, length(y)),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{full}{fitted values from a regression of the outcome on the full set of covariates.}

\item{reduced}{fitted values from a regression of the fitted values from the full regression on the reduced set of covariates.}

\item{y}{the outcome.}

\item{folds}{the folds for hypothesis testing}

\item{type}{which parameter are you estimating (defaults to \code{r_squared}, for R-squared-based variable importance)?}

\item{x}{the covariates, only used if \code{ipc_weights} are entered (defaults to \code{NULL}).}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes unobserved).}

\item{ipc_weights}{weights for inverse probability of coarsening (e.g., inverse weights from a two-phase sample) weighted estimation.}

\item{ipc_fit_type}{if "external", then use \code{ipc_eif_preds}; if "SL", fit a SuperLearner to determine the correction to the efficient influence function}

\item{ipc_eif_preds}{if \code{ipc_fit_type = "external"}, the fitted values from a regression of the full-data EIF on the fully observed covariates/outcome; otherwise, not used.}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}

\item{...}{other arguments to SuperLearner, if \code{ipc_fit_type = "SL"}.}
}
\value{
The estimated variable importance for the given group of left-out covariates.
}
\description{
Compute nonparametric estimates of the chosen variable importance parameter, with a correction for using data-adaptive techniques to estimate the conditional means only if necessary.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
