% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.predictiveness_measure.R
\name{estimate.predictiveness_measure}
\alias{estimate.predictiveness_measure}
\title{Obtain a Point Estimate and Efficient Influence Function Estimate for a Given Predictiveness Measure}
\usage{
\method{estimate}{predictiveness_measure}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"predictiveness_measure"}}

\item{...}{other arguments to type-specific predictiveness measures (currently unused)}
}
\value{
A list with the point estimate, naive point estimate (for ANOVA only),
  estimated EIF, and the predictions for coarsened data EIF (for coarsened data settings only)
}
\description{
Obtain a Point Estimate and Efficient Influence Function Estimate for a Given Predictiveness Measure
}
