% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_eif_projection}
\alias{estimate_eif_projection}
\title{Estimate projection of EIF on fully-observed variables}
\usage{
estimate_eif_projection(obs_grad, C, Z, ipc_fit_type, ...)
}
\arguments{
\item{obs_grad}{the estimated (observed) EIF}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes
unobserved).}

\item{Z}{a matrix (or data.frame) with the fully-observed variables}

\item{...}{other arguments to the estimation tool, see "See also".}
}
\value{
the projection of the EIF onto the fully-observed variables
}
\description{
Estimate projection of EIF on fully-observed variables
}
