% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_type_predictiveness.R
\name{estimate_type_predictiveness}
\alias{estimate_type_predictiveness}
\title{Estimate Predictiveness Given a Type}
\usage{
estimate_type_predictiveness(arg_lst, type)
}
\arguments{
\item{arg_lst}{a list of arguments; from, e.g., \code{predictiveness_measure}}

\item{type}{the type of predictiveness, e.g., \code{"r_squared"}}
}
\description{
Estimate the specified type of predictiveness
}
