% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.npvi.R
\name{format.npvi}
\alias{format.npvi}
\title{Format a \code{npvi} object}
\usage{
\method{format}{npvi}(x, ...)
}
\arguments{
\item{x}{the \code{npvi} object of interest.}

\item{...}{other options, see the generic \code{format} function.}
}
\description{
Nicely formats the output from a \code{npvi} object for printing.
}

