% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.predictiveness_measure.R
\name{format.predictiveness_measure}
\alias{format.predictiveness_measure}
\title{Format a \code{predictiveness_measure} object}
\usage{
\method{format}{predictiveness_measure}(x, ...)
}
\arguments{
\item{x}{the \code{predictiveness_measure} object of interest.}

\item{...}{other options, see the generic \code{format} function.}
}
\description{
Nicely formats the output from a \code{predictiveness_measure} object for printing.
}
