% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.vim.R
\name{format.vim}
\alias{format.vim}
\title{Format a \code{vim} object}
\usage{
\method{format}{vim}(x, ...)
}
\arguments{
\item{x}{the \code{vim} object of interest.}

\item{...}{other options, see the generic \code{format} function.}
}
\description{
Nicely formats the output from a \code{vim} object for printing.
}
