% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate test data}
\usage{
generateData(n, p = 2)
}
\arguments{
\item{n}{the sample size}

\item{p}{the number of covariates. Defaults to 2, and many more will result in increased computation time for \code{np}.}
}
\description{
Generates data used in the low-dimensional simulations.
}
\examples{
generateData(100)
}

