% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sampled_split_predictions.R
\name{get_cv_sl_folds}
\alias{get_cv_sl_folds}
\title{Get a numeric vector with cross-validation fold IDs from CV.SuperLearner}
\usage{
get_cv_sl_folds(cv_sl_folds)
}
\arguments{
\item{cv_sl_folds}{The folds from a call to \code{CV.SuperLearner}; a list.}
}
\value{
A numeric vector with the fold IDs.
}
\description{
Get a numeric vector with cross-validation fold IDs from CV.SuperLearner
}
