% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_set}
\alias{get_test_set}
\title{Return test-set only data}
\usage{
get_test_set(arg_lst, k)

get_test_set(arg_lst, k)
}
\arguments{
\item{arg_lst}{a list of estimates, data, etc.}

\item{k}{the index of interest}
}
\value{
the test-set only data

the test-set only data
}
\description{
Return test-set only data

Return test-set only data
}
