% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lco.R
\name{lco}
\alias{lco}
\title{Leave Covariate(s) Out Nonparametric Variable Importance}
\usage{
lco(f1, f2, data, j = 1, alpha = 0.05, type = "mgcv", ...)
}
\arguments{
\item{f1}{the regression function to estimate for the full fit,
in the usual \code{lm} format.}

\item{f2}{the regression function to estimate for the reduced fit,
in the usual \code{lm} format.}

\item{j}{the covariate(s) to calculate variable importance for,
defaults to 1.}

\item{alpha}{the level to compute the confidence interval at.
Defaults to 0.05, corresponding to a 95\% confidence interval.}

\item{type}{the nonparametric estimation tool to use, defaults to
\code{mgcv}. Partial strings can be used, if they are unique identifiers of the tool.}

\item{...}{other arguments to the estimation tool, see "See also".}

\item{dat}{the dataset. Must be in \eqn{n x (p+1)}, where the last
column is \eqn{Y}.}
}
\value{
An object of class \code{npvi}. See Details for more information.
}
\description{
Compute estimates and confidence intervals for the
nonparametric variable importance parameter of interest.
}
\details{
See the paper by Williamson, Carone, and Simon for more
details on the mathematics behind this function, and the validity
of the confidence intervals.
In the interest of transparency, we return most of the calculations
within the \code{npvi} object. This results in a list containing:
\itemize{
 \item{call}{ the call to \code{lco}}
 \item{full.f}{ the model for the full data}
 \item{red.f}{ the model for the reduced data}
 \item{data}{ the data used by the function}
 \item{j}{ the column(s) to calculate variable importance for}
 \item{type}{ the method to estimate the regressions}
 \item{full.fit}{ the fitted values of the chosen method fit to the full data}
 \item{red.fit}{ the fitted values of the chosen method fit to the reduced data}
 \item{est}{ the estimated variable importance}
 \item{se}{ the standard error for the estimated variable importance}
 \item{ci}{ the \eqn{(1-\alpha) x 100}\% confidence interval for the variable importance}
 \item{full.mod}{ the object returned by the estimation procedure for the full data regression}
 \item{red.mod}{ the object returned by the estimation procedure for the reduced data regression}
 \item{alpha}{ the level, for confidence interval calculation}
}
}
\examples{
## generate the data
testdat <- generateData(n = 100, p = 2)

## get the estimate
est <- lco(y ~ V1 + V2, y ~ V1, data = testdat, j = 2, type = "np", regtype = "ll", cktertype = "epanechnikov")

}
\seealso{
\code{\link{np}} for specific usage of the nonparametric kernel smoothing methods,
\code{\link{mgcv}} for specific usage of the GAM methods, and
\code{\link{loess}} for local polynomial fitting.
}

