% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_kfold}
\alias{make_kfold}
\title{Turn folds from 2K-fold cross-fitting into individual K-fold folds}
\usage{
make_kfold(
  cross_fitting_folds,
  sample_splitting_folds = rep(1, length(unique(cross_fitting_folds))),
  C = rep(1, length(cross_fitting_folds))
)

make_kfold(
  cross_fitting_folds,
  sample_splitting_folds = rep(1, length(unique(cross_fitting_folds))),
  C = rep(1, length(cross_fitting_folds))
)
}
\arguments{
\item{cross_fitting_folds}{the vector of cross-fitting folds}

\item{sample_splitting_folds}{the sample splitting folds}

\item{C}{vector of whether or not we measured the observation in phase 2}
}
\value{
the two sets of testing folds for K-fold cross-fitting

the two sets of testing folds for K-fold cross-fitting
}
\description{
Turn folds from 2K-fold cross-fitting into individual K-fold folds

Turn folds from 2K-fold cross-fitting into individual K-fold folds
}
