% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_anova.R
\name{measure_anova}
\alias{measure_anova}
\title{Estimate ANOVA
Compute nonparametric estimate of the ANOVA decomposition.}
\usage{
measure_anova(
  full,
  reduced,
  y,
  x = NULL,
  C = rep(1, length(y)),
  ipc_weights = rep(1, length(y)),
  ipc_fit_type = "external",
  ipc_eif_preds = rep(1, length(y)),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{full}{fitted values from a regression function of the outcome on the full set of covariates.}

\item{reduced}{fitted values from a regression on the reduced set of covariates.}

\item{x}{the covariates, only used if \code{ipc_weights} are entered (defaults to \code{NULL}).}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes unobserved).}

\item{ipc_weights}{weights for inverse probability of coarsening (e.g., inverse weights from a two-phase sample) weighted estimation.}

\item{ipc_fit_type}{if "external", then use \code{ipc_eif_preds}; if "SL", fit a SuperLearner to determine the correction to the efficient influence function}

\item{ipc_eif_preds}{if \code{ipc_fit_type = "external"}, the fitted values from a regression of the full-data EIF on the fully observed covariates/outcome; otherwise, not used.}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}

\item{...}{other arguments to SuperLearner, if \code{ipc_fit_type = "SL"}.}
}
\value{
A named list of: (1) the estimated ANOVA (based on a one-step correction) of the fitted regression functions; (2) the estimated influence function; (3) the naive ANOVA estimate; and (4) the IPC EIF predictions.
}
\description{
Estimate ANOVA
Compute nonparametric estimate of the ANOVA decomposition.
}
