% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_auc.R
\name{measure_auc}
\alias{measure_auc}
\title{Estimate area under the receiver operating characteristic curve (AUC)}
\usage{
measure_auc(
  fitted_values,
  y,
  x = NULL,
  C = rep(1, length(y)),
  ipc_weights = rep(1, length(y)),
  ipc_fit_type = "external",
  ipc_eif_preds = rep(1, length(y)),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{x}{the covariates, only used if \code{ipc_weights} are entered (defaults to \code{NULL}).}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes unobserved).}

\item{ipc_weights}{weights for inverse probability of coarsening (e.g., inverse weights from a two-phase sample) weighted estimation.}

\item{ipc_fit_type}{if "external", then use \code{ipc_eif_preds}; if "SL", fit a SuperLearner to determine the correction to the efficient influence function}

\item{ipc_eif_preds}{if \code{ipc_fit_type = "external"}, the fitted values from a regression of the full-data EIF on the fully observed covariates/outcome; otherwise, not used.}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}

\item{...}{other arguments to SuperLearner, if \code{ipc_fit_type = "SL"}.}
}
\value{
A named list of: (1) the estimated AUC of the fitted regression function; (2) the estimated influence function; and (3) the IPC EIF predictions.
}
\description{
Compute nonparametric estimate of AUC.
}
