% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_cross_entropy.R
\name{measure_cross_entropy}
\alias{measure_cross_entropy}
\title{Estimate the cross-entropy}
\usage{
measure_cross_entropy(
  fitted_values,
  y,
  weights = rep(1, length(y)),
  na.rm = FALSE
)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{y}{the outcome.}

\item{weights}{weights (IPW, etc.).}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}
}
\value{
A named list of: (1) the estimated cross-entropy of the fitted regression function, and (2) the estimated influence function.
}
\description{
Compute nonparametric estimate of cross-entropy.
}
