% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_deviance.R
\name{measure_deviance}
\alias{measure_deviance}
\title{Estimate the deviance}
\usage{
measure_deviance(fitted_values, y, weights = rep(1, length(y)), na.rm = FALSE)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{y}{the outcome.}

\item{weights}{weights (IPW, etc.).}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}
}
\value{
A named list of: (1) the estimated deviance of the fitted regression function, and (2) the estimated influence function.
}
\description{
Compute nonparametric estimate of deviance.
}
