% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_r_squared.R
\name{measure_r_squared}
\alias{measure_r_squared}
\title{Estimate R-squared
Compute nonparametric estimate of R-squared.}
\usage{
measure_r_squared(fitted_values, y, weights = rep(1, length(y)), na.rm = FALSE)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{y}{the outcome.}

\item{weights}{weights (IPW, etc.).}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}
}
\value{
A named list of: (1) the estimated R-squared of the fitted regression function, and (2) the estimated influence function.
}
\description{
Estimate R-squared
Compute nonparametric estimate of R-squared.
}
