% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npvi.R
\docType{package}
\name{npvi}
\alias{npvi}
\alias{npvi-package}
\title{npvi: Calculate leave covariate(s) out (LCO)
nonparametric variable importance}
\description{
The npvi package provides one major function: lco.
This function calculates an estimate of the variable importance
parameter of interest developed by Williamson, Carone, and Simon.
}
\section{npvi Functions}{

The function lco computes the estimates, standard error estimates,
and confidence intervals. It is an object of class "npvi",
and has its own print method.
}

