% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.npvi.R
\name{plot.npvi}
\alias{plot.npvi}
\title{Plot \code{npvi} objects}
\usage{
\method{plot}{npvi}(x, y, ...)
}
\arguments{
\item{x}{an \code{npvi} object created from a call to \code{combine}.}

\item{y}{a vector of names, given in the same order as the list.}

\item{...}{other options, see the generic \code{plot} function.}
}
\description{
Plot the estimates and standard errors for a set of \code{npvi} objects.
}

