% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vim.R
\name{plot.vim}
\alias{plot.vim}
\title{Plot \code{vim} objects}
\usage{
\method{plot}{vim}(x, y, ...)
}
\arguments{
\item{x}{an \code{vim} object.}

\item{y}{a vector of names, given in the same order as the estimates in the \code{vim} object.}

\item{...}{other options, see the generic \code{plot} function.}
}
\description{
Plot the estimates and standard errors for a set of \code{vim} objects.
}
