% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveness_measure.R
\name{predictiveness_measure}
\alias{predictiveness_measure}
\title{Construct a Predictiveness Measure}
\usage{
predictiveness_measure(
  type = character(),
  y = numeric(),
  a = NULL,
  fitted_values = numeric(),
  full_y = NULL,
  C = numeric(),
  Z = NULL,
  ipc_weights = numeric(),
  ipc_fit_type = character(),
  ipc_eif_preds = numeric(),
  ipc_est_type = character(),
  scale = character(),
  na.rm = logical(),
  ...
)
}
\arguments{
\item{type}{the measure of interest (e.g., "accuracy", "auc", "r_squared")}

\item{y}{the outcome of interest}

\item{a}{the exposure of interest (only used if \code{type = "average_value"})}

\item{fitted_values}{fitted values from a regression function using the
observed data (may be within a specified fold, for cross-fitted estimates).}

\item{full_y}{the observed outcome (not used, defaults to \code{NULL}).}

\item{C}{the indicator of coarsening (1 denotes observed, 0 denotes
unobserved).}

\item{Z}{either \code{NULL} (if no coarsening) or a matrix-like object
containing the fully observed data.}

\item{ipc_weights}{weights for inverse probability of coarsening (IPC)
(e.g., inverse weights from a two-phase sample) weighted estimation.
Assumed to be already inverted.
(i.e., ipc_weights = 1 / [estimated probability weights]).}

\item{ipc_fit_type}{if "external", then use \code{ipc_eif_preds}; if "SL",
fit a SuperLearner to determine the IPC correction to the efficient
influence function.}

\item{ipc_eif_preds}{if \code{ipc_fit_type = "external"}, the fitted values
from a regression of the full-data EIF on the fully observed
covariates/outcome; otherwise, not used.}

\item{ipc_est_type}{IPC correction, either \code{"ipw"} (for classical
inverse probability weighting) or \code{"aipw"} (for augmented inverse
probability weighting; the default).}

\item{scale}{if doing an IPC correction, then the scale that the correction
should be computed on (e.g., "identity"; or "logit" to logit-transform,
apply the correction, and back-transform).}

\item{na.rm}{logical; should \code{NA}s be removed in computation?
(defaults to \code{FALSE})}

\item{...}{other arguments to SuperLearner, if \code{ipc_fit_type = "SL"}.}
}
\value{
An object of class \code{"predictiveness_measure"}, with the following
 attributes:
}
\description{
Construct a Predictiveness Measure
}
