% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npvi.R
\name{print.npvi}
\alias{print.npvi}
\title{Print a \code{npvi} object}
\usage{
\method{print}{npvi}(x, ...)
}
\arguments{
\item{x}{the \code{npvi} object of interest.}

\item{...}{other options, see the generic \code{print} function.}
}
\description{
Prints out the table of estimates, confidence intervals, and standard errors for a \code{npvi} object.
}

