% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predictiveness_measure.R
\name{print.predictiveness_measure}
\alias{print.predictiveness_measure}
\title{Print \code{predictiveness_measure} objects}
\usage{
\method{print}{predictiveness_measure}(x, ...)
}
\arguments{
\item{x}{the \code{predictiveness_measure} object of interest.}

\item{...}{other options, see the generic \code{print} function.}
}
\description{
Prints out a table of the point estimate and standard error for a \code{predictiveness_measure} object.
}
