% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vim.R
\name{print.vim}
\alias{print.vim}
\title{Print a \code{vim} object}
\usage{
\method{print}{vim}(x, ...)
}
\arguments{
\item{x}{the \code{vim} object of interest.}

\item{...}{other options, see the generic \code{print} function.}
}
\description{
Prints out the table of estimates, confidence intervals, and standard errors for a \code{vim} object.
}
