% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_arg_lst}
\alias{process_arg_lst}
\title{Process argument list for Super Learner estimation of the EIF}
\usage{
process_arg_lst(arg_lst)
}
\arguments{
\item{arg_lst}{the list of arguments for Super Learner}
}
\value{
a list of modified arguments for EIF estimation
}
\description{
Process argument list for Super Learner estimation of the EIF
}
