% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_update.R
\name{risk_update}
\alias{risk_update}
\title{Estimate the influence function for an estimator of risk}
\usage{
risk_update(fitted_values, y, weights = rep(1, length(y)),
  type = "r_squared", na.rm = FALSE)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{y}{the outcome.}

\item{weights}{weights for the computed influence curve (e.g., inverse probability weights for coarsened-at-random settings)}

\item{type}{which risk parameter are you estimating (defaults to \code{r_squared}, for the $R^2$)?}

\item{na.rm}{logical; should NAs be removed in computation? (defaults to \code{FALSE})}
}
\value{
The influence function values for the given group of left-out covariates.
}
\description{
Estimate the influence function for the given risk parameter.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
