% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{variableImportance}
\alias{variableImportance}
\title{Estimate variable importance using a one-step estimator-based approach}
\usage{
variableImportance(full, reduced, y, n = length(y), standardized = TRUE,
  two_phase = FALSE)
}
\arguments{
\item{full}{fitted values from a regression of the outcome on the full set of covariates.}

\item{reduced}{fitted values from a regression either (1) of the outcome on the reduced set of covariates, or (2) of the fitted values from the full regression on the reduced set of covariates.}

\item{y}{the outcome.}

\item{n}{the sample size.}

\item{standardized}{logical; should we estimate the standardized parameter? (defaults to \code{TRUE})}

\item{two_phase}{logical; did the data come from a two-phase study? (defaults to \code{FALSE})}
}
\value{
The estimated variable importance for the given group of left-out covariates.
}
\description{
Compute nonparametric estimates of the variable importance parameter interpreted as the proportion of variability explained by including a group of covariates in the estimation technique.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
