% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportanceCI.R
\name{variableImportanceCI}
\alias{variableImportanceCI}
\title{Confidence intervals for variable importance}
\usage{
variableImportanceCI(est, se, n, level = 0.95)
}
\arguments{
\item{est}{estimate of variable importance from a call to \code{variableImportance}.}

\item{se}{estimate of the standard error of \code{est}, from a call to \code{variableImportanceSE}}

\item{n}{the sample size.}

\item{level}{confidence interval type (defaults to 0.95).}
}
\value{
The estimated variable importance for the given group of left-out covariates.
}
\description{
Compute confidence intervals for the estimate of the variable importance parameter interpreted as the proportion of variability explained by including a group of covariates in the estimation technique.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
\examples{


}

