% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportanceIC.R
\name{variableImportanceIC}
\alias{variableImportanceIC}
\title{Estimate the influence function}
\usage{
variableImportanceIC(full, reduced, y, standardized = TRUE)
}
\arguments{
\item{full}{fitted values from a regression of the outcome on the full set of covariates.}

\item{reduced}{fitted values from a regression either (1) of the outcome on the reduced set of covariates, or (2) of the fitted values from the full regression on the reduced set of covariates.}

\item{y}{the outcome.}

\item{standardized}{logical; should we estimate the standardized parameter? (defaults to \code{TRUE})}
}
\value{
The influence function values for the given group of left-out covariates.
}
\description{
Compute the value of the influence function for the given group of left-out covariates.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
\examples{


}

