% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.R
\name{vimp}
\alias{vimp}
\title{Variable Importance}
\usage{
vimp(mod, measures = list(auc), task, data)
}
\arguments{
\item{mod}{An object of class randomForest, as that created by the function randomForest with option keep.inbag = TRUE}

\item{measures}{List of performance measure(s) of mlr to evaluate. Default is auc only.}

\item{task}{Learning task created by the function makeClassifTask or makeRegrTask of mlr.}
}
\value{
Returns a dataframe with a column for each desired measure.
}
\description{
With the help of this function the permutation variable importance for random forests 
can be created for any measure that is available in the mlr package.
}
\examples{
library(mlr)
library(randomForest)

# Classification
data = getTaskData(iris.task)
mod = randomForest(Species ~., data = data, ntree = 100, keep.inbag = TRUE)
results = vimp(mod, measures = list(mmce, multiclass.au1u, multiclass.brier), task = iris.task, data = data)

}

