% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_ci.R
\name{vimp_ci}
\alias{vimp_ci}
\title{Confidence intervals for variable importance}
\usage{
vimp_ci(est, se, scale = "identity", level = 0.95, truncate = TRUE)
}
\arguments{
\item{est}{estimate of variable importance, e.g., from a call to \code{vimp_point_est}.}

\item{se}{estimate of the standard error of \code{est}, e.g., from a call to \code{vimp_se}.}

\item{scale}{scale to compute interval estimate on (defaults to "identity": compute Wald-type CI).}

\item{level}{confidence interval type (defaults to 0.95).}

\item{truncate}{truncate CIs to have lower limit at (or above) zero?}
}
\value{
The Wald-based confidence interval for the true importance of the given group of left-out covariates.
}
\description{
Compute confidence intervals for the true variable importance parameter.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
