% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_hypothesis_test.R
\name{vimp_hypothesis_test}
\alias{vimp_hypothesis_test}
\title{Perform a hypothesis test against the null hypothesis of zero importance}
\usage{
vimp_hypothesis_test(full, reduced, y, type = "r_squared",
  level = 0.05, na.rm = FALSE)
}
\arguments{
\item{full}{fitted values from a regression of the outcome on the full set of covariates.}

\item{reduced}{fitted values from a regression either (1) of the outcome on the reduced set of covariates, or (2) of the fitted values from the full regression on the reduced set of covariates.}

\item{y}{the outcome.}

\item{type}{which parameter are you estimating (defaults to \code{anova}, for ANOVA-based variable importance)?}

\item{level}{the desired type I error rate (defaults to 0.05).}

\item{na.rm}{logical; should NAs be removed in computation? (defaults to \code{FALSE})}
}
\value{
\code{TRUE} if the null hypothesis is rejected (i.e., if the confidence intervals do not overlap); otherwise, \code{FALSE}.
}
\description{
Perform a hypothesis test against the null hypothesis of zero importance by: 
(i) for a user-specified level \eqn{\alpha}, compute a \eqn{(1 - \alpha)\times 100}\% confidence interval around the risk for both the full and reduced regression functions;
(ii) if the intervals do not overlap, reject the null hypothesis.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
