% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_auc.R
\name{measure_auc}
\alias{measure_auc}
\title{Estimate area under the receiver operating characteristic curve (AUC)}
\usage{
measure_auc(fitted_values, y, weights = rep(1, length(y)), na.rm = FALSE)
}
\arguments{
\item{fitted_values}{fitted values from a regression function.}

\item{y}{the outcome.}

\item{weights}{weights (IPW, etc.).}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}
}
\value{
A named list of: (1) the estimated AUC of the fitted regression function, and (2) the estimated influence function.
}
\description{
Compute nonparametric estimate of AUC.
}
