% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npvi.R
\name{print.npvi}
\alias{print.npvi}
\title{Print a \code{npvi} object}
\usage{
\method{print}{npvi}(obj, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{obj}{the \code{npvi} object of interest.}

\item{digits}{the number of digits to display results to. Defaults to the maximum of 3 and the "digits" option minus 3.}
}
\description{
Prints out the table of estimates, confidence intervals, and standard errors for a \code{npvi} object.
}

