% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_sl}
\alias{run_sl}
\title{Run a Super Learner for the provided subset of features}
\usage{
run_sl(Y, X, V, SL.library, s, folds, ...)
}
\arguments{
\item{Y}{the outcome}

\item{X}{the covariates}

\item{V}{the number of folds}

\item{SL.library}{the library of candidate learners}

\item{s}{the subset of interest}

\item{folds}{the CV folds}

\item{...}{other arguments to Super Learner}
}
\value{
a list of length V, with the results of predicting on the hold-out data for each v in 1 through V
}
\description{
Run a Super Learner for the provided subset of features
}
\keyword{internal}
