% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportanceTMLE.R
\name{variableImportanceTMLE}
\alias{variableImportanceTMLE}
\title{Estimate variable importance using a TMLE-based approach}
\usage{
variableImportanceTMLE(full, reduced, y, x, s, lib, tol = .Machine$double.eps)
}
\arguments{
\item{full}{fitted values from an initial regression of the outcome on the full set of covariates.}

\item{reduced}{fitted values from an initial regression either (1) of the outcome on the reduced set of covariates, or (2) of the fitted values from the full regression on the reduced set of covariates.}

\item{y}{the outcome variable}

\item{x}{the covariates (ordered)}

\item{s}{the indices for variable importance}

\item{lib}{the library of candidate learners passed to SuperLearner}

\item{tol}{the tolerance level for convergence}
}
\value{
The estimated variable importance for the given group of left-out covariates.
}
\description{
Compute nonparametric estimates of the variable importance parameter interpreted as the proportion of variability explained by including a group of covariates in the estimation technique.
}
\details{
This differs from estimates returned by variableImportance() in that the procedure used to target the estimate to the parameter of interest is based on a TMLE approach.
}
