% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_se.R
\name{vimp_se}
\alias{vimp_se}
\title{Estimate standard errors}
\usage{
vimp_se(eif, n = length(eif), na.rm = FALSE)
}
\arguments{
\item{eif}{the estimated efficient influence function.}

\item{n}{the sample size.}

\item{na.rm}{logical; should NA's be removed in computation? 
(defaults to \code{FALSE}).}
}
\value{
The standard error for the estimated variable importance for the 
  given group of left-out covariates.
}
\description{
Compute standard error estimates for estimates of variable importance.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
  details on the mathematics behind this function and the definition of the 
  parameter of interest.
}
