% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_se.R
\name{vimp_se}
\alias{vimp_se}
\title{Estimate standard errors}
\usage{
vimp_se(update, n = length(update), na.rm = FALSE)
}
\arguments{
\item{update}{the influence curve-based update}

\item{n}{the sample size}

\item{na.rm}{logical; should NA's be removed in computation? (defaults to \code{FALSE})}
}
\value{
The standard error for the estimated variable importance for the given group of left-out covariates.
}
\description{
Compute standard error estimates for estimates of the variable importance parameter interpreted as the proportion of variability explained by including a group of covariates in the estimation technique.
}
\details{
See the paper by Williamson, Gilbert, Simon, and Carone for more
details on the mathematics behind this function and the definition of the parameter of interest.
}
